requirements.txt
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
auto-gptq @ https://github.com/PanQiWei/AutoGPTQ/releases/download/v0.2.2/auto_gptq-0.2.2+cu117-cp310-cp310-linux_x86_64.whl#sha256=89a71e3fd0476be214b2720f8723a2789427900af070b88ec126fbb0cf1bf5c7
beautifulsoup4==4.12.2
bitsandbytes==0.39.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
colorama==0.4.6
contourpy==1.1.0
cpm-kernels==1.0.11
cycler==0.11.0
dataclasses-json==0.5.9
datasets==2.13.1
dbus-python==1.2.16
dill==0.3.6
diskcache==5.6.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
einops==0.6.1
exceptiongroup==1.1.1
exllama @ https://github.com/jllllll/exllama/releases/download/0.0.4/exllama-0.0.4+cu117-cp310-cp310-linux_x86_64.whl#sha256=b0ea456f02a29ad390d295f988d1b8a11ee4835b21e6a81b00b9060c50ae0e2d
fastapi==0.98.0
ffmpy==0.3.0
filelock==3.12.2
flexgen==0.1.7
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.33.1
gradio_client==0.2.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
icetk==0.0.7
idna==2.8
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.228
langchainplus-sdk==0.0.20
linkify-it-py==2.0.2
lit==16.0.6
llama-cpp-python==0.1.66
loralib==0.1.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.1
packaging==23.1
pandas==2.0.2
peft @ git+https://github.com/huggingface/peft@03eb378eb914fbee709ff7c86ba5b1d033b89524
Pillow==9.5.0
platformdirs==3.8.0
psutil==5.9.5
PuLP==2.7.0
pyarrow==12.0.1
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.36.0
pyparsing==3.1.0
pyrsistent==0.19.3
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.4.2
rouge==1.0.1
safetensors==0.3.1
scipy==1.11.1
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.14.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.22.0
virtualenv==20.23.1
websockets==11.0.3
xxhash==3.2.0
yarl==1.9.2
